+++ week03/django01/posts/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
-from .models import Post
+from .models import *
 
 admin.site.register(Post)
+admin.site.register(Profile)
 
 # Register your models here.
